requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cachetools==5.3.2
google-generativeai==0.3.2  # For Gemini API